requirements.txt
google-adk>=0.1.0
google-cloud-aiplatform>=1.40.0
python-dotenv>=1.0.0
rich>=13.0.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
